requirements.txt
# Core Dependencies
pydantic

# LlamaIndex - let pip resolve compatible versions
llama-index-core
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-vector-stores-chroma

# Vector Database
chromadb

# Agent Framework (optional for Phase 2+)
# pydantic-ai>=0.0.13  # Uncomment when needed for Phase 2

# Utilities
python-dotenv
click
dateparser
pyyaml
opentelemetry-api
opentelemetry-sdk

# Testing
pytest
pytest-asyncio
